JFR-47: dropped the prefer for method names. The prefers has been replaced by a priority, which is
a) more precise: it defines exactly which combination of productions is
disallowed, as opposed to the prefer, which just prefers a production
over everything else.

b) more performant.

GenericStaticMethod has been renamed to GenericMethod and the prefix
of the GenericMethod is now an AmbName, not a TypeName. This
represents the syntactical ambiguity better.



git-svn-id: https://svn.strategoxt.org/repos/StrategoXT/java-front/trunk@15083 ba653a0a-7fb7-0310-a048-973b18a224e4