D    syn/stratego-v1.5/AmbEmbedded-Java-15.sdf D    syn/stratego-v1.5/AmbEmbedded-Java-15-quotations.sdf
Removed, part of java-java.

A    syn/stratego-v1.5/languages
A    syn/stratego-v1.5/languages/java
A    syn/stratego-v1.5/languages/java/EmbeddedJava.sdf
M    syn/stratego-v1.5/Makefile.am

Renovation of the Java embedding stuff. The preferred embedding is now
no longer using a prefixed version of Java, but a mix version of Java.

The preferred definition for embedding Java is now
EmbeddedJavaMix.def. This module is parameterized with two symbols:

  1) The context of Java.
  2) The expression non-terminal of the host language.

If you use StrategoRenamed, then the import can be something like
this:

  languages/java/EmbeddedJavaMix[Java StrategoTerm]

However, we will soon have a StrategoMix as well. In that case, the
import will be:

  languages/java/EmbeddedJavaMix[Java Term[[Stratego]]]

where Stratego is imported as:

  languages/stratego/StrategoMix[Stratego]

The advantage of having EmbeddedJavaMix is that you can easily embed
multiple languages, even if more than one of these languages involve
Java. Suppose that you have a language Foo that extends Java:

-------------------------------------
module Foo
imports
  languages/java/JavaMix[Foo]

exports
   ...
-------------------------------------

and Foo is embedded in Stratego:

-------------------------------------
module StrategoFoo
imports
  languages/stratego/StrategoMix[Stratego]
  Foo

exports
  ...
-------------------------------------

In this situation the composer of the embeddings can decide if he (she
is unlikely) wants to support the variables and anti-quotations in Foo
quotations and if Foo constructs should be supported in Java
quotations.

If this should be allowed, then EmbeddedJavaMix can be imported with
the parameters [Foo Term[[Stratego]]]. If this should not be allowed,
then a different first parameter can be chosen.




git-svn-id: https://svn.strategoxt.org/repos/StrategoXT/java-front/trunk@16394 ba653a0a-7fb7-0310-a048-973b18a224e4